deprecate hotshot-stake-table (#3182)
* deprecate hotshot-stake-table

* fix dev-node

* fix mock_ledger

* wrap `compute_stake_table_commitment` with error & fix CI

* quick lint

+++ sequencer/src/genesis.rs
@@ -15,7 +15,7 @@ use vbs::version::Version;
 /// Initial configuration of an Espresso stake table.
 #[derive(Clone, Debug, Deserialize, Serialize, PartialEq, Eq)]
 pub struct StakeTableConfig {
-    pub capacity: u64,
+    pub capacity: usize,
 }
 
 /// An L1 block from which an Espresso chain should start syncing.
